chore(dist): ignore the dist files but still commit them
Frankly, there surely should be a better way than committing your dist/
in the repo. I have not found one yet, but I am also not really sure
it's possible. Here's an alternative: add them in the gitignore so all
the cli tools will ignore them (rg, fzf, telescope, etc) but for commit
them each time a build is made. So it's impossible to _not_ have them
present in a commit whenever there's a change and `npm run build` is run.