feat(local): write a cli to run the code locally
This will serve as a local-testing suite and a way for people that don't
want to use GitHub Actions to also get the functionalities.

This simplifies the testing process as well, as the need to build/push
the dist folder is reserved for testing specific logic about GitHub
Action's context.

The entire environment should be mocked locally for transparent usage.

fix #36