feat(noop): implement a custom http client
This adds a custom http.Client that responds to Do(). It noops some
requests, and forward others.

This makes the whole noop'ing really simple since we can just set the
custom client at the star of the execution and implement the nooping
cases in the noop package, instead of having to do if/else everywhere.