feat!(config): change default path to `.ln-config.yaml`
By default, the GitHub Action Token cannot write to `.github`. This was
documented in https://github.com/nobe4/action-ln/pull/207.

Given that I really want to be able to self-manage the `ln-config`, I am
changing its default location from `.github/` to the root directory of
the repo. This is a breaking change, but easily fixable, so I don't
expect any problem with it.