fix(config): pull the `To` file from the head branch
This check first if the `To` file exists on the defined head branch, if
that's the case, it's the one we should try to update. If not, then we
can use the ref/default branch.

This will ensure that if the `From` changes whilst a PR is open, it will
not fail to update.

It introduces some additional complexity, but shouldn't be too bad. One
thing to change in the future is where the head branch name is stored.
It might go into the config at some point.

Fix #211