feat(file): always report `From.Ref`
If `From.Ref` is not specified in the configuration, it is fetched
during the configuration population. This ensures that each call to the
API is explicit about the used Ref. It also allows each link to report
the branch used on the PR, even if it's the repo's default.

Fix #214